Refactor pattern handling
Introduce functions to improve the pattern handling. A private module
variable is used to store the patterns for the tags. There are init,
get patterns and get pattern functions. With the init function it is
possible to explicitly initialize the patterns before they are
requested. The get patterns functions return the tag to pattern dict and
initialize the pattern dict if it hasn't be initialized explicitly. The
get pattern functions allow to retrieve a pattern for a tag.